requirements.txt
# Файл requirements.txt

# Официальная библиотека для работы с OpenAI API
openai==0.27.6

# Библиотеки для работы с LangChain
langchain==0.0.213

# Для загрузки переменных окружения из .env файлов
python-dotenv>=1.0.0

# Для работы с HTTP запросами
requests>=2.32.0

# Для работы с регулярными выражениями
regex>=2023.12.18

# Для работы с векторными базами данных
faiss-cpu==1.9.0.post1

beautifulsoup4==4.12.3

uvicorn
fastapi

python-telegram-bot==21.10

